refactored some types
