--- packages/landingpage-new/types/commitPerMonth.ts
@@ -1,1 +1,0 @@
-export type CommitsPerMonth = Record<string, number>; // e.g. Dec 21 => 5
