--- packages/landingpage-new/types/githubActivity.ts
@@ -1,6 +1,0 @@
-export interface GithubActivity {
-    days: number[];
-    total: number;
-    week: number;
-    date?: Date;
-}
