--- packages/landingpage-new/types/githubContributor.ts
@@ -1,13 +1,0 @@
-import { UserTypes } from './githubUserTypes';
-
-export type GithubContributor = {
-  id: number;
-  login: string;
-  html_url: string;
-  avatar_url: string;
-  type: UserTypes;
-};
-
-export type GithubCommitAuthor = {
-  author: GithubContributor;
-};
