--- packages/landingpage-new/types/githubUserTypes.ts
@@ -1,4 +1,0 @@
-export enum UserTypes {
-  USER = 'User',
-  BOT = 'Bot',
-}
